+++ tests/features/di/test_di_main.py
@@ -1,3 +1,4 @@
+import pytest
 from starlette.requests import Request
 
 from lihil import Payload, Route, Text